User.password null on login gives 500 response
Currently `bcrypt.compare` throws an error because the hash argument is required. It's passed in as null/undefined. This results in a 500 when it's not really the server's fault.

Some kind of action is needed on the user's part - a 4xx response. I'm not sure if this is exactly the right solution / error response. Maybe it should throw `password_reset_required` but I don't see this error message anywhere else in this file so I'm not sure if this plugin has that concept.